pyproject.toml
[tool.poetry]
name = "The Finance Crew"
version = "0.1.0"
description = "Ai Crew for Brad"
authors = ["Emilio <emiliosanalvarez07@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "0.27.0"
crewai_tools = "0.1.7"
firebase_admin = "6.5.0"
Flask = "3.0.3"
flask_pymongo = "2.3.0"
groq = "0.5.0"
langchain = "0.1.16"
langchain_anthropic = "0.1.8"
langchain_community = "0.0.32"
langchain_voyageai = "0.1.0"
pydantic = "2.7.0"
python-dotenv = "1.0.0"
Requests = "2.31.0"
sec_api = "1.0.17"
unstructured = "0.13.2"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"